requirements.txt
genanki>=0.13.0
markdown>=3.5.0
PyYAML>=6.0
python-slugify>=8.0.0
Pillow
